Incde Steven changes
So I've merged in Steven changes from 5/9 with the exception of the list_item.xml changes as they dont look right. I've also updated the implementation to be able to update the recycleview based on location (clicking on a building). Still need to fix how the plant icons update though...*rubs chin*