Adjustments on Plant Spec Page
I adjusted some textviews on the plant spec page. In addition, I pulled in a 3rd party library to help format(transform) the images on the plant spec page to be all the same size and have rounded corners. Include this in your gradle for it to work:
implementation 'jp.wasabeef:picasso-transformations:2.2.1'